azure-fonctions/requirements.txt
azure-functions
requests
bs4
python-dateutil
